fix bugs

+++ src/Stats.hpp
@@ -76,7 +76,7 @@ class Stats
             }
             else {
                 double delta = B.mean() - A.mean();
-                double X_n = A.n() + B.n();
+                unsigned long X_n = A.n() + B.n();
                 //double X_mean = A.mean() + delta*(B.n()/X_n);
                 double X_mean = (A.n()*A.mean() + B.n()*B.mean()) / X_n;
                 double X_M2 = A.M2() + B.M2() + delta*delta*A.n()*B.n()/X_n;
@@ -91,7 +91,7 @@ class Stats
             }
         }
 
-        double n() const { return _n; }
+        unsigned long n() const { return _n; }
         double mean() const { return _mean; }
         double M2() const { return _M2; }
         double variance() const { return _M2/(_n-1); }
